feat: Add dynamic padding adjustments and enhance layout state handling
- Introduced `updatePadding` method in `MapCameraState` for dynamic map padding updates.
- Refactored `topAppBarHeight` type from `Float` to `Int` for consistency in layout parameters.
- Improved `MapCameraEffect` to dynamically calculate padding based on UI state changes.
- Enhanced `MapTopAppBar` with height reporting and integrated `onGloballyPositioned` for real-time layout updates.

+++ feature/map/src/androidMain/kotlin/me/matsumo/onenavi/feature/map/MapViewModel.kt
@@ -250,7 +250,7 @@ private class UiEventDelegate(
         )
     }
 
-    private fun handleTopAppBarHeightChanged(height: Float) {
+    private fun handleTopAppBarHeightChanged(height: Int) {
         uiState.value = uiState.value.copy(topAppBarHeight = height)
     }
